pipe for season selector dates
