+++ src/app/components/shared/selectors/season-selector/season-selector.component.html
@@ -9,8 +9,8 @@
     <app-circle-slider [startValue]="this.start" [endValue]="this.end" (newStart)="onSeasonStartChange($event)" (newEnd)="onSeasonEndChange($event)" (doneSelecting)="onSeasonDoneSelecting()"></app-circle-slider>
     <!-- <app-date-range></app-date-range> -->
     <div class="dates">
-      <div><b>Season Start Day </b> <span *ngIf="start"> • {{ dayOfYearFormat(start, 'short') }}</span></div>
+      <div><b>Season Start Day </b> <span *ngIf="start"> • {{ this.start | shortDateSeason}}</span></div>
       <button mat-icon-button	 (click)="swap()" matTooltip="Swap season start and end date"><mat-icon>swap_vert</mat-icon></button	>
-      <div><b>Season End Day </b> <span *ngIf="end"> • {{ dayOfYearFormat(end, 'short') }}</span></div>
+      <div><b>Season End Day </b> <span *ngIf="end"> • {{ end | shortDateSeason }}</span></div>
     </div>
 </div>
