+++ src/app/components/shared/selectors/season-selector/season-selector.module.ts
@@ -9,6 +9,7 @@ import { SeasonSelectorComponent } from './season-selector.component';
 import { CircleSliderModule} from '@components/shared/selectors/circle-slider/circle-slider.module';
 import { DateRangeModule } from '../date-range/date-range.module';
 import { MatButtonModule } from '@angular/material/button';
+import { PipesModule } from '@pipes';
 @NgModule({
   declarations: [SeasonSelectorComponent],
   imports: [
@@ -20,7 +21,8 @@ import { MatButtonModule } from '@angular/material/button';
     MatSliderModule,
     CircleSliderModule,
     DateRangeModule,
-    MatButtonModule
+    MatButtonModule,
+    PipesModule
   ],
   exports: [SeasonSelectorComponent]
 })
